Additional Logging
Previously when there was some sort of failure, an error message was printed
encouraging the review of a log file.  That log file was then immediately
recycled as part of container destruction preventing anything useful being
gleaned from it.  This change actually prints the log file when there is a
failure.